added new GUI related classes in ui

+++ src/main/model/AdditionalOptions.java
@@ -2,8 +2,7 @@ package model;
 
 public class AdditionalOptions extends MenuItem {
 
-    public AdditionalOptions(String name, Integer price) {
+    public AdditionalOptions(String name, Double price) {
         super(name, price);
     }
-
 }
